# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"AppController_getHello_v1","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/v1/auth/verify-email":{"post":{"operationId":"AuthController_verifyEmail_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailDto"}}}},"responses":{"200":{"description":"Email verified successfully"}},"summary":"Verify email with OTP","tags":["Auth"]}},"/api/v1/auth/signup":{"post":{"operationId":"AuthController_userSignUp_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignUpDto"}}}},"responses":{"201":{"description":"User registered successfully"}},"summary":"User Sign Up","tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_signIn_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful"}},"summary":"User / Admin Login","tags":["Auth"]}},"/api/v1/auth/change-password":{"post":{"operationId":"AuthController_changePassword_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"201":{"description":"Password changed successfully"}},"security":[{"access-token":[]}],"summary":"Change Password","tags":["Auth"]}},"/api/v1/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"201":{"description":"OTP sent successfully"}},"summary":"Forgot Password","tags":["Auth"]}},"/api/v1/auth/verify-otp":{"post":{"operationId":"AuthController_verifyOtp_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpDto"}}}},"responses":{"200":{"description":"OTP verified successfully"}},"summary":"Verify OTP for Password Reset","tags":["Auth"]}},"/api/v1/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"Password reset successfully"}},"summary":"Reset Password","tags":["Auth"]}},"/api/v1/auth/refresh-token":{"post":{"operationId":"AuthController_refreshToken_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"Token refreshed successfully"}},"summary":"Refresh access token","tags":["Auth"]}},"/api/v1/auth/me":{"get":{"operationId":"AuthController_getMe_v1","parameters":[],"responses":{"200":{"description":"User profile fetched successfully"}},"security":[{"bearer":[]}],"summary":"Get current authenticated user","tags":["Auth"]}},"/api/v1/auth/profile":{"patch":{"operationId":"AuthController_updateProfile_v1","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":"Profile updated successfully"}},"security":[{"access-token":[]}],"summary":"Update user profile","tags":["Auth"]}}},"info":{"title":"NestJS Prisma Template API","description":"A production-ready NestJS + Prisma API template.\n\n**Base URL:** `http://localhost:{port}/api/v1`\n\n**Auth:** Use the `POST /api/v1/auth/login` endpoint to obtain a Bearer token, then click **Authorize** above.","version":"1.0.0","contact":{}},"tags":[{"name":"Auth","description":"Authentication & authorization endpoints"},{"name":"User","description":"User management endpoints"}],"servers":[{"url":"http://localhost:3000","description":"Local Development"}],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"Enter your JWT access token"}},"schemas":{"VerifyEmailDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"otp":{"type":"string","example":"123456"}},"required":["email","otp"]},"UserSignUpDto":{"type":"object","properties":{"userName":{"type":"string","example":"johndoe"},"email":{"type":"string","example":"user@gmail.com"},"contactNumber":{"type":"string","example":"01700000000"},"password":{"type":"string","example":"12345678"}},"required":["userName","email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@gmail.com"},"password":{"type":"string","example":"12345678"}},"required":["email","password"]},"ChangePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string","example":"OldPassword@123","description":"Current password"},"newPassword":{"type":"string","example":"NewPassword@123","description":"New password (min 6 chars)"}},"required":["oldPassword","newPassword"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"user@gmail.com"}},"required":["email"]},"VerifyOtpDto":{"type":"object","properties":{"email":{"type":"string","example":"user@gmail.com","description":"Email address"},"otp":{"type":"string","example":"123456","description":"6-digit OTP code"}},"required":["email","otp"]},"ResetPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"user@gmail.com","description":"Email address"},"otp":{"type":"string","example":"123456","description":"6-digit OTP code"},"newPassword":{"type":"string","example":"NewPassword@123","description":"New password (min 6 chars)"}},"required":["email","otp","newPassword"]},"RefreshTokenDto":{"type":"object","properties":{"userId":{"type":"string","example":"userId"},"refreshToken":{"type":"string","example":"refresh Token"}},"required":["userId","refreshToken"]},"UpdateProfileDto":{"type":"object","properties":{"userName":{"type":"string","example":"John Doe"},"contactNumber":{"type":"string","example":"+1234567890"},"file":{"type":"string","format":"binary","description":"Profile image file"}}}}}}